dockerized_labs/a02_crypto_failures_2025_lab/requirements.txt
Flask==3.1.0
pycryptodome==3.18.0
cryptography==39.0.1
